move the rest of the command code into the commands
